🔧 chore(interfaces): update solidity version to 0.8.4 in IAddressRegistry.sol, IEcosystemReserve.sol, and IGovernance.sol The solidity version in the interface files has been updated to 0.8.4 to align with the desired version for the project. This ensures compatibility and consistency across the codebase.
🔧 chore(interfaces): update solidity version to 0.8.4 in IInvestorsNFT and IRandToken interfaces
The solidity version in the IInvestorsNFT and IRandToken interfaces has been updated to 0.8.4 to align with the desired version for the project. This update ensures compatibility and consistency with the rest of the codebase.

🔧 chore(IStakedRand.sol): update solidity version to 0.8.4 and format code for better readability
💡 The changes were made to update the Solidity version to 0.8.4 and format the code to improve readability. Updating the Solidity version ensures compatibility with the latest features and enhancements of the language. Formatting the code improves maintainability and makes it easier for developers to understand and work with the contract interface.

🔄 chore(interfaces): update solidity version to 0.8.4 in IStakedToken and IVestingControllerERC721 interfaces
The solidity version in the IStakedToken and IVestingControllerERC721 interfaces has been updated to 0.8.4 to ensure compatibility with the latest version of the Solidity compiler. This update allows the interfaces to take advantage of the new features and improvements introduced in Solidity 0.8.4.