Add SignatureVerification_V2 and VestingControllerERC721_V2 to README.md
This commit adds the files SignatureVerification_V2.md and VestingControllerERC721_V2.md to README.md in the docs folder. The SignatureVerification_V2 file contains information on the events and parameters for the SignatureUsed event. The VestingControllerERC721_V2 file contains information on the methods and parameters for interacting with the VestingControllerERC721_V2 contract.